Delete also the file when deleting the document in the DB

+++ demo-storage/main.py
@@ -86,7 +86,11 @@ def debug_main():
 @app.route('/__debug__/delete/<key>')
 def debug_delete(key):
     files_collection = get_files_collection()
+    filename = get_file_path(key)
+    if not filename:
+        abort(404)
     files_collection.delete_one({"_id": ObjectId(key)})
+    os.remove(os.path.join(app.config['UPLOAD_FOLDER'], filename))
     return redirect(url_for('debug_main'))
 
 if __name__ == '__main__':
